requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
facenet-pytorch==2.5.2
torch
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
opencv-python-headless==4.8.0.76
scipy==1.11.2
python-dotenv==1.0.0
Pillow==10.0.1
python-multipart